Reorganize fetch.
This changes the behaviour of fancypath files.
Previously, a from-disk file would have the prefix
verified before checking if the file existed in
memory, branching off to the from-disk branch at
this point. This meant that for in-memory fiels,
we would have to validate the path prefix first.

This has been changed so that all files are tested
for in-memory presence. If it does not exist, we
check the prefix and potentially from-disk
handling, then to be followed by the old 404
handling. This also saves us a goto.